# api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Damumed КМИС API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus];

export const PatientStatus = {
  active: "active",
  discharged: "discharged",
  scheduled: "scheduled",
} as const;

export interface Patient {
  id: number;
  fullName: string;
  birthDate: string;
  gender: string;
  iinNumber: string;
  diagnosis: string;
  /** @nullable */
  ward?: string | null;
  admissionDate: string;
  /** @nullable */
  dischargeDate?: string | null;
  status: PatientStatus;
  attendingDoctor: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePatientBody {
  fullName: string;
  birthDate: string;
  gender: string;
  iinNumber: string;
  diagnosis: string;
  /** @nullable */
  ward?: string | null;
  admissionDate: string;
  attendingDoctor: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
}

export interface UpdatePatientBody {
  fullName?: string;
  diagnosis?: string;
  /** @nullable */
  ward?: string | null;
  /** @nullable */
  dischargeDate?: string | null;
  status?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
}

export type MedicalFormFormType =
  (typeof MedicalFormFormType)[keyof typeof MedicalFormFormType];

export const MedicalFormFormType = {
  primary_exam: "primary_exam",
  daily_diary: "daily_diary",
  epicrisis: "epicrisis",
  consultation: "consultation",
} as const;

export type MedicalFormStatus =
  (typeof MedicalFormStatus)[keyof typeof MedicalFormStatus];

export const MedicalFormStatus = {
  draft: "draft",
  submitted: "submitted",
} as const;

export interface MedicalForm {
  id: number;
  patientId: number;
  formType: MedicalFormFormType;
  /** @nullable */
  complaints?: string | null;
  /** @nullable */
  anamnesis?: string | null;
  /** @nullable */
  objectiveStatus?: string | null;
  /** @nullable */
  diagnosis?: string | null;
  /** @nullable */
  treatment?: string | null;
  /** @nullable */
  recommendations?: string | null;
  /** @nullable */
  doctorNotes?: string | null;
  status: MedicalFormStatus;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  submittedBy?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateFormBodyFormType =
  (typeof CreateFormBodyFormType)[keyof typeof CreateFormBodyFormType];

export const CreateFormBodyFormType = {
  primary_exam: "primary_exam",
  daily_diary: "daily_diary",
  epicrisis: "epicrisis",
  consultation: "consultation",
} as const;

export interface CreateFormBody {
  patientId: number;
  formType: CreateFormBodyFormType;
  /** @nullable */
  complaints?: string | null;
  /** @nullable */
  anamnesis?: string | null;
  /** @nullable */
  objectiveStatus?: string | null;
  /** @nullable */
  diagnosis?: string | null;
  /** @nullable */
  treatment?: string | null;
  /** @nullable */
  recommendations?: string | null;
  /** @nullable */
  doctorNotes?: string | null;
}

export interface UpdateFormBody {
  /** @nullable */
  complaints?: string | null;
  /** @nullable */
  anamnesis?: string | null;
  /** @nullable */
  objectiveStatus?: string | null;
  /** @nullable */
  diagnosis?: string | null;
  /** @nullable */
  treatment?: string | null;
  /** @nullable */
  recommendations?: string | null;
  /** @nullable */
  doctorNotes?: string | null;
}

export type AppointmentSpecialistType =
  (typeof AppointmentSpecialistType)[keyof typeof AppointmentSpecialistType];

export const AppointmentSpecialistType = {
  doctor: "doctor",
  physiotherapy: "physiotherapy",
  massage: "massage",
  psychologist: "psychologist",
  speech_therapy: "speech_therapy",
} as const;

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: number;
  patientId: number;
  patientName: string;
  specialistType: AppointmentSpecialistType;
  specialistName: string;
  scheduledDate: string;
  scheduledTime: string;
  duration: number;
  status: AppointmentStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export type CreateAppointmentBodySpecialistType =
  (typeof CreateAppointmentBodySpecialistType)[keyof typeof CreateAppointmentBodySpecialistType];

export const CreateAppointmentBodySpecialistType = {
  doctor: "doctor",
  physiotherapy: "physiotherapy",
  massage: "massage",
  psychologist: "psychologist",
  speech_therapy: "speech_therapy",
} as const;

export interface CreateAppointmentBody {
  patientId: number;
  specialistType: CreateAppointmentBodySpecialistType;
  specialistName: string;
  scheduledDate: string;
  scheduledTime: string;
  duration: number;
  /** @nullable */
  notes?: string | null;
}

export type UpdateAppointmentBodyStatus =
  (typeof UpdateAppointmentBodyStatus)[keyof typeof UpdateAppointmentBodyStatus];

export const UpdateAppointmentBodyStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateAppointmentBody {
  status?: UpdateAppointmentBodyStatus;
  /** @nullable */
  notes?: string | null;
}

export interface GenerateScheduleBody {
  patientId: number;
  startDate: string;
  procedures?: string[];
}

export interface DashboardSummary {
  totalPatients: number;
  activePatients: number;
  todayAppointments: number;
  completedToday: number;
  pendingForms: number;
  scheduledDischarges: number;
}

export type ListPatientsParams = {
  search?: string;
  status?: string;
};

export type ListAppointmentsParams = {
  patientId?: number;
  date?: string;
  specialistType?: string;
};