# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Damumed КМИС API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
...
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        birthDate:
          type: string
        gender:
          type: string
        iinNumber:
          type: string
        diagnosis:
          type: string
        ward:
          type: ["string", "null"]
        admissionDate:
          type: string
        dischargeDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, discharged, scheduled]
        attendingDoctor:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - fullName
        - birthDate
        - gender
        - iinNumber
        - diagnosis
        - admissionDate
        - status
        - attendingDoctor
        - createdAt
        - updatedAt
    CreatePatientBody:
      type: object
      properties:
        fullName:
          type: string
        birthDate:
          type: string
        gender:
          type: string
        iinNumber:
          type: string
        diagnosis:
          type: string
        ward:
          type: ["string", "null"]
        admissionDate:
          type: string
        attendingDoctor:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required:
        - fullName
        - birthDate
        - gender
        - iinNumber
        - diagnosis
        - admissionDate
        - attendingDoctor
...